requirements.txt
lightgbm>=4.0.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
python-dateutil>=2.8.0
